invoice/app/requirements.txt
Flask
mysql-connector
setuptools